[refactor] refactoring all code and making code review requests
