--- Customer.API/Program.cs
@@ -1,38 +1,0 @@
-using Customer.AppServices.Services;
-using Customer.AppServices.Services.Interface;
-using Customer.DomainServices.Services;
-using Customer.DomainServices.Services.Interfaces;
-using Customer.Infrastructure.Data.Context;
-using FluentValidation;
-using FluentValidation.AspNetCore;
-using Microsoft.EntityFrameworkCore;
-using System.Reflection;
-
-var builder = WebApplication.CreateBuilder(args);
-
-var connectionString = builder.Configuration.GetConnectionString("DefaultConnection");
-builder.Services.AddDbContext<FeatureContext>(
-    options => options.UseMySql(connectionString, ServerVersion.AutoDetect(connectionString)));
-
-builder.Services.AddControllers();
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-builder.Services.AddTransient<ICustomerService, CustomerService>();
-builder.Services.AddTransient<ICustomerAppService, CustomerAppService>();
-builder.Services.AddFluentValidationAutoValidation();
-builder.Services.AddValidatorsFromAssembly(Assembly.Load("Customer.AppServices"));
-builder.Services.AddAutoMapper(Assembly.Load("Customer.AppServices"));
-
-var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
-
-app.UseAuthorization();
-
-app.MapControllers();
-
-app.Run();
