--- Customer.AppModels/Customer.AppModels.csproj
@@ -1,10 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-	  <NoWarn>8618</NoWarn>
-  </PropertyGroup>
-
-</Project>
