--- Customer.AppModels/Dtos/CustomerResult.cs
@@ -1,56 +1,0 @@
-﻿namespace Customer.AppModels.Dtos
-{
-    public class CustomerResult
-    {
-        public CustomerResult(
-            long id,
-            string fullName,
-            string email,
-            string emailConfirmation,
-            string cpf,
-            string cellphone,
-            DateTime dateOfBirth,
-            bool emailSms,
-            bool whatsapp,
-            string country,
-            string city,
-            string postalCode,
-            string address,
-            int number)
-        {
-            Id = id;
-            FullName = fullName;
-            Email = email;
-            EmailConfirmation = emailConfirmation;
-            Cpf = cpf;
-            Cellphone = cellphone;
-            DateOfBirth = dateOfBirth;
-            EmailSms = emailSms;
-            Whatsapp = whatsapp;
-            Cpf = cpf;
-            Country = country;
-            City = city;
-            PostalCode = postalCode;
-            Address = address;
-            Number = number;
-        }
-
-        protected CustomerResult() { }
-
-        public long Id { get; set; }
-        public string FullName { get; set; }
-        public string Email { get; set; }
-        public string EmailConfirmation { get; set; }
-        public string Cpf { get; set; }
-        public string Cellphone { get; set; }
-        public string Country { get; set; }
-        public string City { get; set; }
-        public string Address { get; set; }
-        public string PostalCode { get; set; }
-        public bool EmailSms { get; set; }
-        public bool Whatsapp { get; set; }
-        public int Number { get; set; }
-        public DateTime DateOfBirth { get; set; }
-    }
-}
-
