--- Customer.AppServices/Profile/CustomerProfile.cs
@@ -1,16 +1,0 @@
-﻿using AutoMapper;
-using Customer.AppModels.Dtos;
-using Customer.DomainModels.Models;
-
-namespace Customer.AppServices.Mappings
-{
-    public class CustomerProfile : Profile
-    {
-        public CustomerProfile()
-        {
-            CreateMap<CreateCustomerDto, CustomerModel>();
-            CreateMap<CustomerModel, CustomerResult>();
-            CreateMap<UpdateCustomerDto, CustomerModel>();
-        }
-    }
-}
