--- Customer.DomainModels/Models/BaseModel.cs
@@ -1,7 +1,0 @@
-﻿namespace Customer.DomainModels.Models
-{
-    public abstract class BaseModel
-    {
-        public long Id { get; set; }
-    }
-}
