--- Customer.DomainServices/Customer.DomainServices.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Customer.Infrastructure.Data\Customer.Infrastructure.Data.csproj" />
-  </ItemGroup>
-
-</Project>
