--- Customer.sln
@@ -1,74 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32825.248
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Customer.AppServices", "Customer.AppServices\Customer.AppServices.csproj", "{70ECEC44-56CD-4D9C-9C17-666EB2C7BD7E}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Customer.DomainServices", "Customer.DomainServices\Customer.DomainServices.csproj", "{1FC26DE9-DB82-40B5-8013-D975CC2FB9FD}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Customer.DomainModels", "Customer.DomainModels\Customer.DomainModels.csproj", "{B79D3418-DBB4-4C43-A992-1EE304851F36}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Customer.API", "Customer.API\Customer.API.csproj", "{76537561-6D84-4F04-B948-E0574369251E}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "API", "API", "{28C1C466-642F-4399-AFD6-3774AADD9520}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Core", "Core", "{7ED4D27A-E11D-474A-92FE-A06D580488A6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Customer.AppModels", "Customer.AppModels\Customer.AppModels.csproj", "{9EE622D2-79E9-48B8-991B-31067B700C5A}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{80D0E9A4-5DF6-45E5-A859-FE7E3F097168}"
-	ProjectSection(SolutionItems) = preProject
-		docker-compose.yml = docker-compose.yml
-	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Infrastructure", "Infrastructure", "{9921A81E-8548-4538-B62F-DB7EB9D60753}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Customer.Infrastructure.Data", "Customer.Infrastructure.Data\Customer.Infrastructure.Data.csproj", "{2161DEB1-FF8E-442E-9DB5-EC504AFC05C1}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{70ECEC44-56CD-4D9C-9C17-666EB2C7BD7E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{70ECEC44-56CD-4D9C-9C17-666EB2C7BD7E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{70ECEC44-56CD-4D9C-9C17-666EB2C7BD7E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{70ECEC44-56CD-4D9C-9C17-666EB2C7BD7E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1FC26DE9-DB82-40B5-8013-D975CC2FB9FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1FC26DE9-DB82-40B5-8013-D975CC2FB9FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1FC26DE9-DB82-40B5-8013-D975CC2FB9FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1FC26DE9-DB82-40B5-8013-D975CC2FB9FD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B79D3418-DBB4-4C43-A992-1EE304851F36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B79D3418-DBB4-4C43-A992-1EE304851F36}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B79D3418-DBB4-4C43-A992-1EE304851F36}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B79D3418-DBB4-4C43-A992-1EE304851F36}.Release|Any CPU.Build.0 = Release|Any CPU
-		{76537561-6D84-4F04-B948-E0574369251E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{76537561-6D84-4F04-B948-E0574369251E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{76537561-6D84-4F04-B948-E0574369251E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{76537561-6D84-4F04-B948-E0574369251E}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9EE622D2-79E9-48B8-991B-31067B700C5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9EE622D2-79E9-48B8-991B-31067B700C5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9EE622D2-79E9-48B8-991B-31067B700C5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9EE622D2-79E9-48B8-991B-31067B700C5A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2161DEB1-FF8E-442E-9DB5-EC504AFC05C1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2161DEB1-FF8E-442E-9DB5-EC504AFC05C1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2161DEB1-FF8E-442E-9DB5-EC504AFC05C1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2161DEB1-FF8E-442E-9DB5-EC504AFC05C1}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{70ECEC44-56CD-4D9C-9C17-666EB2C7BD7E} = {7ED4D27A-E11D-474A-92FE-A06D580488A6}
-		{1FC26DE9-DB82-40B5-8013-D975CC2FB9FD} = {7ED4D27A-E11D-474A-92FE-A06D580488A6}
-		{B79D3418-DBB4-4C43-A992-1EE304851F36} = {7ED4D27A-E11D-474A-92FE-A06D580488A6}
-		{76537561-6D84-4F04-B948-E0574369251E} = {28C1C466-642F-4399-AFD6-3774AADD9520}
-		{9EE622D2-79E9-48B8-991B-31067B700C5A} = {7ED4D27A-E11D-474A-92FE-A06D580488A6}
-		{2161DEB1-FF8E-442E-9DB5-EC504AFC05C1} = {9921A81E-8548-4538-B62F-DB7EB9D60753}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4BD44F92-8664-4418-B498-AAC2D7E9F7A1}
-	EndGlobalSection
-EndGlobal
