+++ DomainModels/Extensions/StringExtensions.cs
@@ -1,0 +1,10 @@
+﻿namespace DomainModels.Extensions
+{
+    public static class StringExtensions
+    {
+        public static string Formatter(this string str)
+        {
+            return str.Trim().Replace(".", "").Replace("-", "");
+        }
+    }
+}
