+++ DomainModels/IEntity.cs
@@ -1,0 +1,7 @@
+﻿namespace DomainModels
+{
+    public interface IEntity
+    {
+        long Id { get; set; }
+    }
+}
